Add Fallback to First Household Record when no HoH is Set
The USPS Cascadia script was failing due to a record not having its head of household
set up. This change adds a fallback to the first household record index when no head
of household is set and emits a warning so that this may be resolved by the study team.