Refactor Ordering Scripts to Use New Utility Modules
This change refactors ordering scripts to use the new utility modules. This eliminates
the weird import structure within the scripts and makes them more logical to follow.
Now script files contain just a main function, which should also make these easier to
work with in a serverless fashion once we get to that point.