Suspend Deliveries When a Participant is Under Study Pause
Prior to this change, whether or not a participant was under a study pause was not taken
into account when sending them kits. Now, participants will not be sent kits while they
are under study pause. The logic flow will skip kit ordering for them and once their
pause has ended kit ordering for the patient will resume.