Fix Bug that Incorrectly Shoved Participants Into Cascadia PDX
Two different bugs caused this issue to be as widespread as it was. First off,
the project name is not included in symptom survey data, so the project name was
not populating correctly and defaulting to PDX. Secondly on enrollment instances
the project name is only present on the index participant (not the head of household),
so when we had a head of household that was not the index participant we had the name
populate incorrectly and default to PDX.

This change breaks out project name finding and mapping into a separate function so that
we can be more confident that we are getting the true project name for a household. Note
that we also no longer default to PDX, we default to a Nonetype project name with a logged
warning that we could not assign a project to the household.