Drop Index so that All Address Fields Populate
A bug previously existed that occasionally dropped last name, email, phone number, and
delivery instructions from certain delivery entries. The root cause of this issue was
that these entries were a mixture of addresses from participants within a household, and
Pandas wasn't allowing the indexes to mix in a way that I thought was logical. Dropping
the index of these addresses allowed the mix and match logic that I desired and eliminated
the issue.

Because of this index change, it was also necessary to update the order id generation function.
This update brings us back in line with the desired id format `YYMMDD_HHID`.