Clean-up S3 Technical Debt
The changes that allowed for a quick deployment of the new S3 upload functionality
were completed quickly and included some technical debt. These changes refactor that
functionality out into utils package and cleanup the script logic flow so that the
code is simpler and easier to follow - eliminating this technical debt.