Ensure Cascadia Orders Have Complete Symptom Survey
This change adds an additional filter for Cascadia participants that ensures they
have a completed symptom survey. The REDCap report is supposed to filter out
records that have incomplete surveys, but it does so inconsistently it seems.
The three most recent issues with unnecessary pickups involved participants with
incomplete symptom surveys, so this change should more effectively filter those
records out of the order sheet.