Upload DE Orders to S3 from Within Script
This change adds command line options to the delivery express order generation
script, which allows the user to specify whether they want the script to save,
upload to S3, or both, the generated order form.

This change also removes the email script and previous S3 upload script as they
are no longer necessary.