Upload USPS Orders from Within Script
This change adds support for uploading USPS orders from within the order
generation script. Based off of the previous commit, it adds command line
options to either save or upload the order form to S3. The USPS form
was not previously uploaded, so there was no need to remove any other files.