Remove Filter on Yesterday's Orders
While testing other changes to the return order script, I realized that the
filter on yesterday's orders introduced a bug. The return order script was
limited to filling in tracking numbers for yesterday's orders, leading to a
backlog in what got return tracking numbers piped in. With the transition to
the new REDCap report, there is no longer a need for this filter, so I am
removing it to also remove the bug from the return tracking script.