Only Resupply Particpants if They've Received Barcodes
A previous version of the resupply function would resupply participants even if they
had zero previous swabs assigned to them simply because this zero swabs was below
the threshold value. This change makes it explicit that participants should have
some number of barcodes before being eligible for a resupply shipment.