Only Send Welcome Kits if a Household is Fully Consented and Enrolled
As the Cascadia study was getting off the ground, we sent welcome kits to a participant if
they were fully consented and enrolled even if the household wasn't. Now that the study is
further along, this creates additional shipments that we can eliminate. This change only
makes a participant eligible for a welcome shipment once their whole household is fully
consented and enrolled, even if that participant was consented and enrolled before that.