Use a Logging File for Config and Restructure Project
This change uses a logging file for configuring the project logger that borrows heavily
from the logging configuration used in [ID3C](https://github.com/seattleflu/id3c/tree/master/lib/id3c/logging). The configuration file is loaded by
the logger prior to each run and propagates down to other modules. Our logging setup
is now less repetitive and easier to parse.

Along with this logging change I also updated the directory structure slightly so that
it played nicer with the logging. We are slightly limited since these are run as scripts
and so have to place utility and logging modules within our path, but overall these changes
make the logging output friendlier and prepares us well for an eventual transition to a
lambda build.