Drop Records with a Duplicated Index Before DE Tracking Import
Prior to this change, the DE tracking number would be piped into every matching
record ID in the DE order report. This implicitly assumed that a participant
would not show up in the report two days in a row and have both rows be populated
by the same order number.

This change drops any rows with a duplicated index, removing this assumption and
making it explicit that we only want to pipe order numbers back to one order per
participant per day. The assumption here is that we want to pipe this number back
to the older order, since this is the one that will have an order number generated
for it already. This should be a much safer assumption and eliminate this bug, which
was causing DE to skip pickups if a participant swabbed two days in a row and both
swab records were populated by this script.