Check Study Pause Status Before Welcome Kit Assignment
The previous version of this script checked if a participant needed a welcome kit
before checking if they were on study pause. This had the effect of assigning a
welcome kit and creating an order for a participant if they were on study pause
should they need a welcome kit. This change moves the study pause check before
the welcome kit check to ensure we don't create any orders for a participant if
they are on study pause.